Moved 'apps' into frontend
