+++ core/frontend/apps/amp/lib/views/amp.hbs
