+++ core/frontend/apps/private-blogging/lib/views/private.hbs
