+++ core/frontend/apps/subscribers/lib/helpers/subscribe_form.js
@@ -2,7 +2,7 @@
 // Usage: `{{subscribe_form}}`
 const _ = require('lodash'),
     // (Less) dirty requires
-    proxy = require('../../../../helpers/proxy'),
+    proxy = require('../../../../../server/helpers/proxy'),
     createFrame = proxy.hbs.handlebars.createFrame,
     templates = proxy.templates,
     urlUtils = proxy.urlUtils,
