+++ core/test/unit/apps/amp/amp_components_spec.js
@@ -1,7 +1,7 @@
 var should = require('should'),
 
 // Stuff we are testing
-    ampComponentsHelper = require('../../../../server/apps/amp/lib/helpers/amp_components');
+    ampComponentsHelper = require('../../../../frontend/apps/amp/lib/helpers/amp_components');
 
 describe('{{amp_components}} helper', function () {
     it('adds script tag for a gif', function () {
