+++ core/test/unit/apps/amp/amp_content_spec.js
@@ -1,7 +1,7 @@
 var should = require('should'),
     rewire = require('rewire'),
     urlUtils = require('../../../../test/utils/urlUtils'),
-    ampContentHelper = rewire('../../../../server/apps/amp/lib/helpers/amp_content');
+    ampContentHelper = rewire('../../../../frontend/apps/amp/lib/helpers/amp_content');
 
 // TODO: Amperize really needs to get stubbed, so we can test returning errors
 // properly and make this test faster!
