+++ core/test/unit/apps/private-blogging/controller_spec.js
@@ -4,7 +4,7 @@ var should = require('should'),
     path = require('path'),
     configUtils = require('../../../utils/configUtils'),
     themes = require('../../../../server/services/themes'),
-    privateController = require('../../../../server/apps/private-blogging/lib/router');
+    privateController = require('../../../../frontend/apps/private-blogging/lib/router');
 
 describe('Private Controller', function () {
     var res, req, defaultPath, hasTemplateStub;
@@ -37,7 +37,7 @@ describe('Private Controller', function () {
             params: {}
         };
 
-        defaultPath = path.join(configUtils.config.get('paths').appRoot, '/core/server/apps/private-blogging/lib/views/private.hbs');
+        defaultPath = path.join(configUtils.config.get('paths').appRoot, '/core/frontend/apps/private-blogging/lib/views/private.hbs');
 
         configUtils.set({
             theme: {
